[BE/feat] 카테고리 삭제 시 기타 카테고리로 일괄 변경 쿼리 구현

+++ backend/src/main/java/ssafy/haruman/domain/category/service/CategoryServiceImpl.java
@@ -13,28 +13,33 @@ import ssafy.haruman.domain.category.dto.response.CategorySimpleResponseDto;
 import ssafy.haruman.domain.category.entity.Category;
 import ssafy.haruman.domain.category.entity.CustomStatus;
 import ssafy.haruman.domain.category.repository.CategoryRepository;
+import ssafy.haruman.domain.challenge.repository.ExpenseRepository;
 import ssafy.haruman.domain.member.entity.Member;
 import ssafy.haruman.global.error.exception.CategoryDuplicateException;
 import ssafy.haruman.global.error.exception.CategoryNotFoundException;
+import ssafy.haruman.global.error.exception.CategoryUnauthorizedException;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final ExpenseRepository expenseRepository;
 
     @Override
     @Transactional
     public CategorySimpleResponseDto createCategory(
             Member member, CategoryCreateRequestDto createDto) {
 
-        Optional<Category> category = categoryRepository.findByName(member, createDto.getName());
+        Optional<Category> category =
+                categoryRepository.findByName(member.getProfile(), createDto.getName());
+
         if (category.isPresent()) {
-            throw new CategoryDuplicateException();
+            throw CategoryDuplicateException.EXCEPTION;
         }
 
         Category createdCategory = Category.builder()
-                .member(member)
+                .profile(member.getProfile())
                 .name(createDto.getName())
                 .isDefault(CustomStatus.CUSTOM)
                 .color(createDto.getColor())
@@ -50,9 +55,10 @@ public class CategoryServiceImpl implements CategoryService {
 
         Category category = findOneCategoryById(updateDto.getCategoryId());
 
-        // TODO 2) 수정 요청 category가 해당 회원의 custom category인지 확인
+        if (!category.getProfile().equals(member.getProfile())) {
+            throw CategoryUnauthorizedException.EXCEPTION;
+        }
 
-        // TODO 3) category 수정
         category.updateCategory(updateDto.getName(), updateDto.getColor());
 
         return CategorySimpleResponseDto.from(category);
@@ -60,24 +66,27 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional
-    public void deleteCategory(Member member, Long categoryId) {
+    public Integer deleteCategory(Member member, Long categoryId) {
 
         Category category = findOneCategoryById(categoryId);
 
-        // TODO 2) 삭제 요청 category가 해당 회원의 custom category인지 확인
+        if (!category.getProfile().equals(member.getProfile())) {
+            throw CategoryUnauthorizedException.EXCEPTION;
+        }
 
-        // TODO 3) 삭제 요청 category에 해당되는 지출 내역 처리 (기타 카테고리로 변경) ** 벌크연산
+        Integer updatedExpenseCnt =
+                expenseRepository.updateCategory(member.getProfile().getId(), category.getId());
 
-        // TODO 4) category 삭제
         categoryRepository.delete(category);
 
+        return updatedExpenseCnt;
     }
 
     @Override
     public List<CategoryDetailResponseDto> selectCategoryList(Member member) {
 
-        // TODO 요청 회원의 custom category + default category 조회
-        List<Category> categoryList = categoryRepository.findAllByProfileAndStatus(member);
+        List<Category> categoryList =
+                categoryRepository.findAllByProfileAndStatus(member.getProfile());
 
         // TODO 요청 회원의 지출 내역에서 가장 많은 상위 N개의 카테고리 조회 (null 제외)
 
@@ -89,8 +98,8 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public List<CategoryDetailResponseDto> selectCustomCategoryList(Member member) {
 
-        // TODO 요청 회원의 custom category 조회
-        List<Category> customCategoryList = categoryRepository.findAllByProfile(member);
+        List<Category> customCategoryList =
+                categoryRepository.findAllByProfile(member.getProfile());
 
         return customCategoryList.stream()
                 .map(CategoryDetailResponseDto::from)
